feat: oauth info 조회 위치 수정, 리뷰 반영
