+++ src/main/kotlin/com/oksusu/susu/user/application/UserFacade.kt
@@ -4,10 +4,10 @@ import com.oksusu.susu.auth.model.AuthUser
 import com.oksusu.susu.config.database.TransactionTemplates
 import com.oksusu.susu.exception.ErrorCode
 import com.oksusu.susu.exception.FailToCreateException
+import com.oksusu.susu.exception.NoAuthorityException
 import com.oksusu.susu.extension.executeWithContext
 import com.oksusu.susu.user.model.request.UpdateUserInfoRequest
 import com.oksusu.susu.user.model.response.UserInfoResponse
-import com.oksusu.susu.user.model.response.UserOAuthInfoResponse
 import org.springframework.stereotype.Service
 
 @Service
@@ -19,7 +19,11 @@ class UserFacade(
         return userService.findByIdOrThrow(user.id).run { UserInfoResponse.from(this) }
     }
 
-    suspend fun updateUserInfo(user: AuthUser, request: UpdateUserInfoRequest): UserInfoResponse {
+    suspend fun updateUserInfo(uid: Long, user: AuthUser, request: UpdateUserInfoRequest): UserInfoResponse {
+        if (uid != user.id) {
+            throw NoAuthorityException(ErrorCode.NO_AUTHORITY_ERROR)
+        }
+
         val user = userService.findByIdOrThrow(user.id)
 
         val updatedUser = txTemplate.writer.executeWithContext {
@@ -31,9 +35,5 @@ class UserFacade(
         } ?: throw FailToCreateException(ErrorCode.FAIL_TO_CREATE_USER_ERROR)
 
         return UserInfoResponse.from(updatedUser)
-    }
-
-    suspend fun getOAuthInfo(user: AuthUser): UserOAuthInfoResponse {
-        return userService.findByIdOrThrow(user.id).run { UserOAuthInfoResponse.from(this) }
     }
 }
